+++ app/props/form-content.props.ts
@@ -1,0 +1,4 @@
+import type { HTMLAttributes } from "vue";
+
+
+export type FormContentProps = Partial<HTMLAttributes>
