+++ components/ui/form/FormContent.vue
@@ -1,11 +1,16 @@
 <template>
-	<div id="form-content" v-bind="$attrs">
+	<div
+		id="form-content"
+		v-bind="$attrs"
+		:class="cn('space-y-1.5', props.class)"
+	>
 		<slot />
 	</div>
 </template>
 
 <script setup lang="ts">
-defineOptions({
-	inheritAttrs: false,
-});
+import type { FormContentProps } from "@/app/props/form-content.props";
+import { cn } from "@/app/utils";
+
+const props = defineProps<FormContentProps>();
 </script>
