+++ pages/index.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import { useHome } from "@/composables/pages/use-home";
+import { useHome } from "@/composables/pages/useHome";
 
 const { msg } = useHome();
 </script>
