+++ tailwind.config.js
@@ -83,6 +83,8 @@ module.exports = {
 				lg: "var(--radius)",
 				md: "calc(var(--radius) - 2px)",
 				sm: "calc(var(--radius) - 4px)",
+            xs: "calc(var(--radius) - 6px)",
+            xxs: "calc(var(--radius) - 8px)",
 			},
 			keyframes: {
 				"accordion-down": {
